+++ packages/api/src/dataProviders/db/rnaSiren/rnaSiren.port.ts
@@ -1,6 +1,6 @@
 import { Rna, Siren } from "dto";
 import { MongoServerError } from "mongodb";
-import { DuplicateIndexError } from "../../../shared/errors/dbErrror/DuplicateIndexError";
+import { buildDuplicateIndexError, isDuplicateError } from "../../../shared/helpers/MongoHelper";
 import RnaSirenEntity from "../../../entities/RnaSirenEntity";
 import MongoRepository from "../../../shared/MongoRepository";
 import { isRna } from "../../../shared/Validators";
@@ -19,8 +19,9 @@ export class RnaSirenPort extends MongoRepository<RnaSirenDbo> {
         try {
             await this.collection.insertOne(RnaSirenAdapter.toDbo(entity));
         } catch (e: unknown) {
-            if (e instanceof MongoServerError && e.code === 11000) { // One or many entities already exist in database but other entities have been saved
-                throw new DuplicateIndexError(e.message);
+            if (e instanceof MongoServerError && isDuplicateError(e)) {
+                // One or many entities already exist in database but other entities have been saved
+                throw buildDuplicateIndexError(e);
             }
             throw e;
         }
@@ -29,13 +30,17 @@ export class RnaSirenPort extends MongoRepository<RnaSirenDbo> {
     async find(query: Rna | Siren) {
         let dbos: RnaSirenDbo[] = [];
         if (isRna(query)) {
-            dbos = await this.collection.find({
-                rna: query
-            }).toArray();
+            dbos = await this.collection
+                .find({
+                    rna: query,
+                })
+                .toArray();
         } else {
-            dbos = await this.collection.find({
-                siren: query
-            }).toArray();
+            dbos = await this.collection
+                .find({
+                    siren: query,
+                })
+                .toArray();
         }
 
         if (!dbos.length) return null;