+++ packages/api/src/modules/search/interfaces/http/SearchController.ts
@@ -11,14 +11,14 @@ export class SearchController extends Controller {
     /**
      * Recherche une association via son rna, siren et nom partiel ou complet
      * @summary Recherche une association via son rna, siren et nom partiel ou complet
-     * @param rna_or_siren Identifiant RNA ou Identifiant Siren
+     * @param input Identifiant RNA ou Identifiant Siren ou Nom d'une association (peut-être encodé via encodeURIComponent())
      */
     @Get("/associations/{input}")
     @Response<HttpErrorInterface>("404", "Aucune association retrouvée", {
         message: "Could match any association with given input : ${input}",
     })
     public async findAssociations(input: string): Promise<{ result: AssociationNameDto[] }> {
-        const result = await searchService.getAssociationsKeys(input);
+        const result = await searchService.getAssociationsKeys(decodeURIComponent(input));
         if (result.length === 0) this.setStatus(204);
         return { result };
     }