+++ packages/api/src/shared/errors/dbErrror/DuplicateIndexError.ts
@@ -1,7 +1,0 @@
-import { ConflictError } from "../httpErrors";
-
-export class DuplicateIndexError extends ConflictError {
-    constructor(message: string, public entity?: unknown) {
-        super(message);
-    }
-}